enable search plugins
