+++ AlcatrazTour/Base.lproj/Main.storyboard
@@ -4,10 +4,10 @@
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="6736"/>
     </dependencies>
     <scenes>
-        <!--Plugin List Main View Controller-->
+        <!--Plugin Table View Controller-->
         <scene sceneID="pMO-Kp-NYn">
             <objects>
-                <tableViewController id="q8b-z6-J4d" customClass="PluginListMainViewController" customModule="AlcatrazTour" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="q8b-z6-J4d" customClass="PluginTableViewController" customModule="AlcatrazTour" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="22" sectionFooterHeight="22" id="bMB-jv-ujm">
                         <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
