--- AlcatrazTour/PluginListSearchResultViewController.swift
@@ -1,24 +1,0 @@
-//
-//  PluginListSearchResultViewController.swift
-//  AlcatrazTour
-//
-//  Created by haranicle on 2015/03/15.
-//  Copyright (c) 2015年 haranicle. All rights reserved.
-//
-
-import UIKit
-import Realm
-
-class PluginListSearchResultViewController: PluginListBaseViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-}
